pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "thesisdatarepo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-storage>=3.0.0",
    "gsutil>=5.36",
    "pymupdf>=1.26.0",
]

[project.scripts]
thesis-gcs-nlp = "thesisdatarepo.cli_run_gcs:main"
thesis-cluster = "thesisdatarepo.cli_analysis:main"

[dependency-groups]
dev = ["pytest>=8.0.0"]
analysis = [
    "sentence-transformers>=3.0.0",
    "umap-learn>=0.5.6",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["maks/thesisdatarepo"]

[tool.pytest.ini_options]
pythonpath = ["maks", "."]
testpaths = ["maks/tests"]
